Version 1.0.07
Fixed the Send command on Windows 95/98/Me so that keys such as {Down}, {Home}, and {PgUp} work properly. [thanks S_R]

Fixed the built-in variable A_ScriptDir so that .aut scripts see it as containing a final backslash (for compatibility). Also changed the .aut to .ahk converter so that this adjustment is made during the conversion. [thanks Rajat]

Fixed "Send, {}}" (send a literal right-brace). [thanks Rajat]

Fixed silent exit when a shifted reserved symbol such as # or ^ was defined as a hotkey. [thanks Rajat]

Fixed ShiftAltTab when it was assigned to a Numpad key. [thanks Rajat]

Changed the following commands so that they don't report runtime errors when given negative parameters: StringLeft, StringRight, StringMid, StringTrimLeft, StringTrimRight, ControlClick, and MouseClick. [thanks Rajat]

Changed the Suspend command so that hotkey subroutines whose first line is "Suspend, on" are not exempt from suspension. [thanks Jay D. Novak]

Changed MsgBox so that it will default to displaying "Press OK to continue" only when called explicitly with zero parameters. [thanks Rajat]

Changed the program's icons to the stylish ones provided by Rajat. It's hard to please everyone with a change like this, but hopefully most of you will consider them to be improvements over the drab originals (which basically just involved me drawing a big purple H).

Improved RegRead, RegWrite, and RegDelete so that they support a key's default value, which is displayed by RegEdit as "(Default)". Also, the root key names can now be abbreviated (e.g. HKLM). [thanks Rajat]

Added command "Menu", which allows you to customize the script's tray icon menu by removing existing items or adding new items. [thanks beardboy & Rajat]