Version 1.0.08
Fixed a small memory leak in Run and RunWait.

Fixed the Start Menu's launch of the main program: with older versions, it's possible that your AutoHotkey.ini script file (if you have one) was placed in the ShellNew folder of your Windows folder. If so, move it from there into the AutoHotkey directory in Program Files (or wherever you installed it).

Fixed the "Edit This Script" menu item, which disabled hotkeys if the script was already open for editing.

Fixed the fact that timers would sometimes make the tray menu and main menu slow to respond.

Fixed GetKeyState when retrieving the physical state of the mouse buttons: it was wrong whenever the mouse hook was installed without the keyboard hook. [thanks skrommel]

Fixed the tilde (~) hotkey prefix so that it works on all keys. [thanks Rajat]

Fixed the fact that the script would silently exit if a blank dynamic variable was used, such as in this example: %EmptyVar% = 100 ; [thanks Rajat]

Changed the tray icon to be closer to the intended green color on Windows 95/98/NT/2000.

Changed one of the types of buffering that was introduced in version 1.0.06 to default to OFF because it does more harm than good in most cases. See #MaxThreadsBuffer for how to re-enable this type of buffering and for an explanation. [thanks jamestr]

Improved #InstallKeybdHook and #InstallMouseHook to accept the optional word FORCE, which avoids the warning dialog that is normally displayed when the hook is already installed by another script. [thanks Rajat]

Improved ControlSend to optionally send directly to the parent window of a control. This is useful for programs such as Winamp 2.x (see Automating Winamp for an example). [thanks Rajat]

Added support for keyboards that have special keys not found on the key list. If your keyboard has such keys, you might be able to make them into hotkeys by using these steps (XP/2000/NT only). [thanks Rajat]

Added NumpadEnter to the key list since it can be made a hotkey on XP/2000/NT. [thanks Rajat]

Added support for hexadecimal integers. Use a 0x prefix to indicate that a number is hex (e.g. 0xF9). Hex integers can be used anywhere that a decimal integer is expected; for example, "Sleep, 0xFF" is equivalent to "Sleep, 255". You can also use SetFormat to make math operations store their results as hex, and to convert between decimal and hex. [thanks Rajat]

Added an option to the Installer that enables drag & drop onto script files. This option is disabled by default since it can lead to accidental launches. [thanks Rajat]

Added Rajat's TextPad syntax file and clip library to the Extras folder, which is accessible via the Start Menu. Use the included script to install these features. Hopefully, similar add-ons for other editors will be developed in the coming months (volunteers are welcome).

Added a "rename" option to the "Menu" command.

Added commands PostMessage/SendMessage.

Added command RunAs, which allows processes to be created under the permissions of a different user. [thanks Jonathan Bennett & Robert Yaklin]

Added command WinSet, which can make a window "always on top" or transparent.