Version 1.0.12
Fixed the script compiler (ahk2exe) so that it will see FileInstall commands even if they are on the same line as an IF command as in this example: IfMsgBox, Yes, FileInstall, ... [thanks Rajat]

Fixed the ExcludeText parameter of all windowing commands. Previously, it didn't do much at all.

Fixed some problems involving rare usages of a key/button as both a hotkey prefix and suffix. In the following example, the "A" key did not repeat when held down. In addition, the ^a hotkey would fire on key-up when it should have fired on key-down:
~a & b::MsgBox "B" was pressed while "A" was held down.
$^a::MsgBox ^a

Fixed non-suppressed prefix keys that are also used as modified suffixes (rare): their up-events were being incorrectly suppressed as in this example:
~!LButton::ToolTip, clicked
~LButton & RButton::MsgBox The right button was clicked while the left button was held down. [thanks Rajat]

Fixed composite hotkeys that use a neutral (non-left/right specific) prefix such as "Control & LShift::". In addition, neutral suffixes are now supported, e.g. "AppsKey & Control::".

Fixed the wrong hotkey firing when a key was physically pressed or held down during a Send operation (very rare even then). This fix only works for hook hotkeys.

Changed for Windows NT4 only: CTRL/ALT/WIN/SHIFT are handled more specifically as left or right rather than neutral, which may improve support for some types of hook hotkeys.

Improved the Hotkey command to optionally toggle a hotkey to its opposite state (enabled or disabled). Some optimizations were also done to this command. [thanks beardboy]

Improved MouseGetPos to optionally retrieve the unique ID number of the window under the mouse cursor. This ID number can be used in the WinTitle parameter of all windowing commands by preceding it with the string ahk_id.

Added command WinGet, which retrieves the unique ID number for a window, or finds all the windows that match the specified criteria.