Version 1.0.16
Possible script-breaking changes: Although 99% of scripts should be unaffected, this top section lists all the changes that might be a problem. I tried to get as many of them into this release as possible to minimize the need for this in the future.

Changed default BatchLines from 10 to 10ms, which makes lengthy operations -- such as loops that have many iterations -- run much faster. The large majority of scripts should continue to function properly. However, if you want 100% compatibility with the prior version of AutoHotkey, add SetBatchLines10 to the top of any script that does not already use the SetBatchLines command.

Changed scripts that use hotkeys, hotstrings, or #Persistent so that #SingleInstance is automatically implied. Use #SingleInstance Off to override this.

Changed the Input command so that ENTER is seen as the more-appropriate linefeed (`n) rather than carriage return (`r). This affects the contents of the OutputVar and the MatchList. Please change your scripts accordingly.

Changed Run/RunWait so that they will launch non-EXE files (such as documents), using the default verb rather than "open". Since the default verb is usually "open", this shouldn't affect most existing scripts.

(normal changes below this point)

Fixed Run/RunWait of .lnk files (shortcuts). [thanks ealerner]

Fixed StringUpper/StringLower so that they work on accented letters and other letters whose ASCII value is above 127.

Fixed Progress so that the W (window width) option works again.

Fixed Shift+Win hotkeys that used the Send command: In some cases, extra shift key events were being generated, which interfered with the ability to use a hotkey two or more times in a row. [thanks Jabberwock]

Fixed ambiguity with `:: when it's used as as part of a hotkey rather than literally in a command. [thanks Rajat]

Fixed InputVars that dynamically resolve to the names of environment variables. Sometimes the value was not being fetched from the environment [thanks Samlii]. Example:
EnvVar = SystemRoot
StringTrimLeft, OutputVar, %EnvVar%, 0

Fixed the ALT key so that if it is itself a hotkey that fires on key-down, keys sent by the Send command will not be modified by the ALT key. [thanks Wade]

Fixed A_LoopFileDir to be blank for files contained in the root folder of a relative-path search.

Fixed the Input command so that having more than 1024 items in the match-list does not cause a memory problem. In addition, activating the Ctrl-Alt-Del security screen no longer causes it to think that CTRL and ALT are stuck down. [thanks frk and beardboy]

Changed the Input command so that Control-Backspace and other modified backspaces are ignored. Only plain backspace is used as an "undo" keystroke.

Changed custom menu items so that they cannot interrupt the OnExit subroutine (as documented). In addition, custom menu items now behave more like hotkeys in terms of buffering and their ability to interrupt other threads.

Increased max size of each variable from 16 to 64 MB (this does not increase the memory used by scripts). This can be made a configurable option if anyone needs it.

It is now possible to use a literal pair of colons by escaping it, e.g. `::

Scan code prefix keys are now possible, e.g. SC118 & Enter::MsgBox. [thanks Rajat]

Improved SplashImage:
- "Keep aspect ratio" can be achieved by specifying -1 for either ZW (object width) or ZH (object height). [thanks MikeG]
- The image is fully redrawn when you drag another window across it (prevents artifacts).
- The image does not "flash" while the user is resizing the window.

Improved ControlClick so that a specific X and Y coordinate within the control can be clicked. [thanks Beastmaster]

Added support for the AutoIt v2 command HideAutoItWin, which does the same thing as Menu, Tray, Icon|NoIcon

Added support for thread priority, which determines whether a thread can interrupt (or be interrupted by) other threads. The following commands support setting the priority: Hotkey, Menu, SetTimer, #Hotstring, and Thread (sets the current thread's priority). In addition, hotstrings can be individually assigned a priority when they are defined.

Added commands SendRaw and ControlSendRaw, which send keystrokes exactly as they appear rather than translating {Enter} to an ENTER keystroke, ^c to Control-C, etc.

Added built-in variables for month name (A_MMMM and A_MMM) and day name (A_DDDD and A_DDD). In addition, A_Year may now be optionally referred to as A_YYYY, A_Mon as A_MM, and A_MDay as A_DD.

Added built-in hotstrings and auto-replace. This is a fairly major feature and due to its variety of uses, it will probably be refined in the coming weeks based on your ideas and feedback.