Version 1.0.17
Fixed Ctrl-Alt-Del security screen and Win+L (lock the computer) to prevent sticking keys. Both were partly broken by v1.0.16. [thanks frk]

Fixed hotstrings for keyboard layouts that use dead keys -- such as acute (&acute;), circumflex (^), and grave (`) -- so that the keyboard will continue to function properly. [thanks /Torben & Jordi]

Fixed auto-replace hotstrings so that end-chars reserved by the Send command, such as !{}, are sent as part of the replacement.

Fixed registry loops so that an entire root key can be recursively searched. [thanks sleepyy35]

Fixed command Suspend, which sometimes permanently disabled all hotstrings. [thanks Loneaussie]

Fixed commands Suspend and Hotkey so that they avoid resetting the state of hotkeys that the user is holding down. [thanks WhiteCloud]

Fixed the Input command so that a match-list contained in a variable will not be ignored if its first character is a comma. [thanks Rajat]

Changed the Input command's MatchList so that two consecutive commas results in a single literal comma. In this example, a single literal comma is included at the end of word1: "word1,,,word2".

Added command "Transform, OutputVar, deref", which expands variable references and escape sequences contained inside other variables. [thanks various]

Added built-in variable A_EndChar, which is the ending character pressed by the user to trigger the most recent non-auto-replace hotstring. If no ending character was required (due to the * option), this variable will be blank.

Added built-in variables A_CaretX and A_CaretY, which are the current X and Y coordinates of the caret (text insertion point).