Version 1.0.18
Fixed sticking modifier keys (CTRL, ALT, SHIFT, and WIN) -- at least in some cases. This should help prevent hook hotkeys from unexpectedly firing when the user isn't even holding down any modifier keys. The fix operates by correcting the hook's modifier state whenever it disagrees with what the OS says it should be.

Fixed mouse button hotkeys: If a script defines LButton as a hotkey, the script's own menus should now operate correctly. Similarly, if a script defines RButton as a hotkey and its subroutine uses "MouseClick, right", the tray icon should now respond correctly to a right-click as long as SetMouseDelay is at the default setting or lower.

Fixed support for diacritic and other characters, which was broken in v1.0.16/17. [thanks jordi]

Fixed some problems with neutral modifier keys (Shift, Control, Alt) that are used as both a prefix and suffix as in this example:
Shift & Capslock::MsgBox You pressed Capslock while holding down the either Shift key.
Shift::MsgBox You released either Shift key without having used it to modify another key.

Fixed "CoordMode Caret" so that it uses its own setting rather than that of "CoordMode ToolTip".

Fixed A_CaretX & Y so that they are always both blank if there is no active window or the caret position could not be determined.

Changed the Input command so that the user can toggle CapsLock/NumLock/ScrollLock during an invisible Input.

Improved the hook's handling of neutral modifier keys (ALT, CTRL, and SHIFT) to make them more specific to which key (left or right) was actually pressed. This is most likely to affect Windows NT4.

Added command "if var [not] between value1 and value2", which checks whether a variable's contents are numerically or alphabetically between two values.

Added command "if var [not] in value1,value2,...", which checks whether a variable's contents match one of the items in a list.

Added built-in variable A_ThisMenuItemPos, which is a number indicating the current position of A_ThisMenuItem within A_ThisMenu.

Added command Process, which performs one of the following operations on a process: checks if it exists; changes its priority; closes it; waits for it to close. [thanks AutoIt Team] In conjunction with this, the Run command has been enhanced to optionally retrieve the Process ID of a newly launched program. Also, a window's Process ID can be retrieved with WinGet.