Version 1.0.19
Fixed hotstrings and the Input command to recognize characters produced with AltGr (Ctrl+Alt), such as the @ symbol. [thanks Tekl, jordi, and /Torben]

Fixed hotstrings so that caseless characters such as digits do not disqualify an abbreviation from being considered uppercase. For example, if "::pt1::pteranodon" is a hotstring and you type PT1 (all caps) the replacement will now be in all caps as it should. In addition, other minor changes to the "case conform" mode have been made.

Fixed the Menu-DeleteAll command so that it doesn't permanently disable any "sub-sub-menus" indirectly owned by a menu. [thanks compuboy_r]

Fixed command "Process Priority" so that the priority level can be contained in a variable. [thanks Rajat]

Fixed commands that turn on the keyboard hook (such as Input) so that they will not re-enable suspended hotkeys. [thanks jordi]

Fixed auto-centering of InputBox, SplashImage, Progress, and SplashTextOn: If the task bar was anchored on the left or top side of the screen, the centering was off by the thickness of the taskbar.

Fixed WinGetText and ControlGetText so that the retrieved text is limited to the maximum size of a variable, which defaults to 64 MB. [thanks hermitress]

Changed BlockInput so that input blocking is momentarily disabled whenever the Send command needs to simulate an ALT keystroke. Since there is a slight chance this will break existing scripts that attempt to send ALT while BlockInput is ON, please check and adjust accordingly.

Improved BlockInput to optionally come on automatically during any of the following commands: Send, SendRaw, MouseMove, MouseClick, and MouseClickDrag. [thanks beardboy]

Slightly improved the monitoring and buffering of hotkeys (especially joystick) during times when the script is displaying a dialog or is uninterruptible.

Added directive #MaxMem, which allows the maximum capacity of each variable to be changed from its default of 64 MB.

Added an option to Run/RunWait that avoids the display of the warning dialog and the termination of the current thread. [thanks compuboy_r and others who have asked]

Added an option to change the background color of a menu via "Menu, MenuName, Color, ColorValue" (has no effect on Windows 95/NT).

Added command KeyWait, which waits for a key or mouse/joystick button to be released or pressed down.

Added command "if var [not] contains MatchList", which is the same as using IfInString/IfNotInString except that multiple search strings are supported (any one of which will cause a match). [thanks whynot]

Added command GUI, which creates and manages windows and controls. Such windows can be used as data entry forms or as custom user interfaces.