Version 1.0.22
Fixed ToolTip failing to appear on Windows XP. [thanks Phi and /Torben]

Fixed the left mouse button when it is used as a hotkey: if it displayed a pop-up menu, the foreground window was seeing the button as stuck-down after the menu was dismissed. [thanks umek]

Improved SetKeyDelay with an optional delay between the down-event and up-event of each keystroke.

Improved GetKeyState, KeyWait, Input, and hotkeys to support an explicit virtual key such as vkFF. This is useful for keys that have no name and that produce no visible character when pressed. Unfortunately, all such keys keys tend to share the same virtual key code (FF); so if you have more than one of them, this method will not be able to tell them apart (except hotkeys, which should use the scNNN method described at the bottom of the key list page).

Improved FileMoveDir to support an optional rename mode, which provides "all or none" behavior. [thanks Tekl].

Added command "WinGet MinMax", which retrieves -1 if the window is minimized, 1 if it is maximized, or 0 if it is neither.

Added "WinGet ProcessName", which retrieves a window's process name. [thanks Wademan]

Added command Drive, which ejects/retracts the tray in a CD or DVD drive, or sets a drive's volume label.

Added command "DriveGet StatusCD", which retrieves the media status of a CD or DVD drive.

Added built-in variables A_ScreenWidth and A_ScreenHeight, which contain the width and height of the primary monitor, in pixels.

Added built-in variable A_AhkVersion, which contains the AutoHotkey version number.

Added command SysGet, which retrieves screen resolution, multi-monitor info, the dimensions of system objects, and other system properties. [thanks Gregory F. Hogg of Hogg&rsquo;s Software]

CHANGES FOR GUI

Fixed Edit controls that have an explicit height (H) that is tall enough to be multi-line: word-wrapping is ON by default, as intended. [thanks Alfredo]

Fixed possible crash (never actually observed) when a window destroys itself, such as via a Quit button.

Fixed "Gui Show" so that the window's previous dimensions are retained if unspecified. [thanks lingoist]

Fixed "Gui Show" so that when called from the tray menu (and perhaps other times), the window will have keyboard focus. [thanks Tekl]

Slightly changed the height of a GroupBox when height is omitted or a row-count is specified. It should now conform better to the current font size. In addition, if no width is specified for a GroupBox, it will be slightly narrower.

Changed "GuiControl Choose" to automatically revert to the ChooseString method if the item-parameter is not numeric.

Improved ListBox to support an optional horizontal scroll bar via the word HScroll in its options.

Improved picture controls to support many additional image formats such as PNG, TIF, and ICO if the OS supports them. If your OS is older than XP (except Windows 95), you can add support by downloading the free Gdi+ extension from Microsoft: www.microsoft.com/downloads/details.aspx?FamilyID=6a63ab9c-df12-4d41-933c-be590feaa05a

Added command "GuiControl Text", which is the same as the set-contents method except for radio buttons and checkboxes, in which case Param3 is treated as the new text/caption even if it equal to -1, 0, or 1. [thanks Rajat]

Added new control type: Tab (SysTabControl32), which is a relatively large control containing multiple pages, each of which contains other controls.