Version 1.0.23
Changed compiled scripts to include a blank version info resource. To customize the version info of your compiled scripts, use a utility such as Resource Hacker (freeware) to edit the AutoHotkeySC.bin file. Note that ResHacker will corrupt compiled scripts, which is why only the AutoHotkeySC.bin file should be edited. [thanks Larl]

Improved the $ hotkey prefix to work on Windows 95/98/Me. It allows a hotkey to send itself, which makes it possible for a hotkey to conditionally perform its original/native-OS function.

Improved PixelGetColor and PixelSearch with an option to retrieve colors in RGB vs. BGR format.

Improved "Menu, Show" to accept an optional position for the menu. [thanks compuboy_r]

Improved the performance of the Hotkey command when it turns on/off a hotkey that is already in the right state. [thanks deguix]

Added sub-commands Style and ExStyle to WinGet, which retrieve a window's style and extended style. [thanks ranomore]

Added built-in variables A_ProgramFiles and A_WinDir for operating systems that might lack the corresponding environment variables.

Added sub-command "WinSet Bottom", which sends a window to the bottom of the stack; that is, beneath all other windows. [thanks AntonyB]

Added sub-command "WinSet, Transparent, Off", which completely turns off transparency on a window that was previously transparent. This may help performance. [thanks dijiyd]

Added sub-command "WinSet TransColor", which can make a window's background transparent on Windows 2000/XP+. This allows the creation of on-screen displays and other visual effects

CHANGES FOR GUI

Fixed "Gui Show": Sometimes the window wouldn't be shown until after the command that followed "Gui Show". [thanks Tekl]

Fixed Picture controls that have a border so that the size of the control doesn't increase slightly each time GuiControl is used to load a new picture. [thanks ILL.1]

Fixed Tab control so that when it is hidden or shown, the controls on its current page are also hidden/shown. In addition, some minor keyboard navigation issues were fixed. [thanks Nemroth]

Changed Tab controls so that enabling or disabling them will also enable/disable their sub-controls.

Changed Tab controls to use Classic Theme appearance on Windows XP or later, which makes their background color match that of their sub-controls. [thanks Gre]

Removed the 32 KB limit on the amount of text that can be contained in an Edit control. Windows 9x will be limited to 64 KB and other operating systems to as much as 4 GB. [thanks Bahri]

Improved screen drawing performance when switching pages in a Tab control.

Improved overall GUI performance, especially for windows with a large number controls.

Added wp+n and hp+n control positioning options, which use the width and height of the previous control.

Added BackgroundTrans option for Text, Picture, and GroupBox controls. It allows other controls behind them to show through.

Added built-in variables A_Gui and A_GuiControl, which contain the GUI window number and Control ID responsible for launching the current thread. [thanks Rajat]

Added sub-command "GuiControl Font", which changes a control's font.

Added control type Slider, which is a sliding bar that the user can move along a vertical or horizontal track.

Added control type Progress, which is a dual-color bar typically used to indicate how much progress has been made toward the completion of an operation.

Added control type Hotkey, which is a box that looks like a single-line edit control but instead accepts a keyboard combination pressed by the user.