Version 1.0.25
NOTE: Although this release has been extensively tested, quite a few low-level changes were made.  It is recommended that mission-critical scripts be re-tested prior to adopting this version.

Fixed file-pattern loops on Windows 95/98/Me so that A_LoopFileTimeModified/Created/Accessed are not blank. [thanks 3tones]

Fixed the following keys on Windows 95/98/Me to allow them to part of hotkey combinations: Delete, Insert, Home, End, PageUp, PageDown, and the arrow keys.

Fixed lines such as "If x = 5, Exit" so that the item after the comma is not recognized as a command. By contrast, "IfEqual, x, 5, Exit" will still recognize "Exit" as a command.

Changed "Transform, OutputVar, Pow" to yield a blank string rather than 0.00 when zero is raised to a negative power.

Changed compiled scripts so that the version resource shows the version number rather than being entirely blank. [thanks g4]

Changed FileAppend so that the file is automatically opened in binary mode if the text to be written contains any carriage return and linefeed pairs (`r`n). In other words, the asterisk option is put into effect automatically. However, specifying the asterisk when appropriate improves performance because the program does not need to scan for `r`n.

Changed file-reading loops so that the output file is not opened if nothing is ever written to it. This happens if the Loop performs zero iterations or if it never uses the FileAppend command. In addition, the file is automatically opened in binary mode if the Loop's first use of FileAppend writes any carriage return and linefeed pairs (`r`n). In other words, the asterisk option is put into effect automatically. These changes might break existing scripts that rely upon: 1) a zero-length file being created by the loop; or 2) the file's modification time being changed by the loop even if it didn't write any text to the file.

Removed runtime validation of numeric parameters (to help performance). Formerly, a command such as "SetKeyDelay %Var%" would produce a runtime error if Var contained something non-numeric. Most commands now treat non-numeric parameters as zero.

Improved the Progress command with options for starting position and having a range other than 0 to 100. [thanks ranomore]

Improved the Send command so that mouse buttons can be held down or released as in this example: Send {LButton down}

Improved the Sort command with an option to eliminate duplicates from the list. Also, the numeric option now treats items as floating point numbers rather than integers. [thanks Rajat]

Improved WIN and ALT hotkeys that use the Send command:
1) Fixed some situations that resulted in an unwanted appearance of the Start Menu.
2) The language bar's hotkey (LAlt+Shift) should no longer get triggered as an unwanted side-effect of the Send command. However, it can still be triggered intentionally with "Send {LAlt down}{Shift}{LAlt up}".
3) ALT hotkeys should no longer activate the menu bar of the active window. If you're using {AltUp}{Alt} anywhere in your scripts to make ALT hotkeys work better, those hotkeys should still work. However, they can be made to behave more consistently by removing {AltUp}{Alt}.

Improved performance of the following: 1) StringReplace "All" uses temporary memory (when appropriate) to greatly speed up the operation. See StringReplace for details and options. 2) Variable concatenation, such as Var = %Var%String, is at least 100 times faster. 3) Assigning to variables when AutoTrim is "On" is faster. 4) Case insensitive string searching is much faster, which affects StringGetPos, IfInString, "if var in <list>", StringReplace, and Input.

Added built-in variable A_LoopFileShortPath to file-pattern loops, which allows the 8.3 short path and name of a file to be retrieved. [thanks Rajat]

Added command FileRead, which reads all of a file's text into a variable (performs better than a file-reading loop).

Added support for expressions in commands that have numeric parameters, such as StringLeft. In addition, IF-statements may use complex expressions by including a parenthesis as the first character, e.g. if (X < Y + 10) and (Color = "Blue"). Finally, Var:=expression has been added to assign the result of an expression to a variable, e.g. Var := 100*X + 1/3. See Expressions for details about operators, string concatenation, and easier access to arrays. [thanks Joost Mulders]

CHANGES FOR GUI

Changed "Gui Show" to launch the GuiSize label only after showing the window and updating its title. [thanks ranomore]

Improved drag-and-drop to report (via ErrorLevel) the number of dropped files.