Version 1.0.25.05
Fixed crash when an expression compared a non-numeric string to the numeric result of an operation. [thanks unnamed.ahk]

Fixed := so that it's not necessary to have a space to its left. [thanks unnamed.ahk]

Fixed string concatenation such as Var = %Var% Test, which was partially broken by v1.0.25. This affected only concatenations that required one or more spaces to be trimmed from the left side due to AutoTrim being on. [thanks namazu(catfish)]

Fixed mouse button hotkeys that use only WIN or ALT as a modifier so that they do not display the Start Menu or activate the active window's menu bar, respectively. This problem occurred only in scripts that were not using the keyboard hook. [thanks Kijuto Riddle]

Fixed scripts that assign the mouse to be AltTab, ShiftAltTab, and other alt-tab actions so that they activate the keyboard hook, which is required for their proper operation. [thanks jonny]

Improved FileSelectFile with a easier multi-select option. Although the old multi-select option will continue to function as before, it has been marked obsolete in the help file due to the difficulty of parsing its results. [thanks 3tones]

Improved StringReplace with an option to store the number of replacements in ErrorLevel.