Version 1.0.25.12
Fixed RunAs so that it can be used more than once. [thanks MikeG]

Fixed "Menu, MyMenu, Standard" to create the menu if it doesn't yet exist. [thanks Mythos]

Fixed "if Var" and "if (expression)" to consider non-numeric strings to be "true", as documented. [thanks Mythos]

Fixed built-in variables so that they can be referenced dynamically even if never referenced directly. Example: A_IPAddress%A_Index%

Changed ErrorLevel so that its previous value is available to commands that give it a new value. Example: StringReplace, OutputVar, ErrorLevel, EndKey:

Changed the linefeed character (`n) to be sent as Enter rather than Ctrl+Enter. This also causes "SendRaw %A_EndChar%" to work correctly when the user presses Enter to complete a hotstring. [thanks Andreas Borutta]