Version 1.0.25.13
Applying options or styles to an existing GUI window avoids showing the window if it is hidden or minimized.

Added the following options to "Gui Show": Minimize, Maximize, Restore, NoActivate, NA, and Hide. [thanks kmel]

Added the following Gui window options: +/-Disabled, +/-AlwaysOnTop, and +/-LastFound. The LastFound option sets the window to be the last found window, allowing commands such as WinSet to operate on it even if it is hidden.