Version 1.0.26
Fixed A_LoopFileLongPath, which was truncating long filenames produced from short ones. [thanks deXter & jonny]

Fixed EnvDiv and "Transform Mod" to allow values between -1.0 and 1.0. [thanks Mythos]

Fixed hotkeys such as "a & b::" to be seen as "~a & b::" if "~a::" also exists as a hotkey, and vice versa. [thanks jonny]

Fixed unwanted interaction between hotkeys. For example, if "Capslock & A" and "MButton" are both hotkeys, clicking the middle button while holding down Capslock will turn CapsLock On/Off as expected.

Changed "MouseClick WheelUp|WheelDown" to ignore SetMouseDelay. This was done to avoid a slow-down of mouse wheel hotkeys, and also because a delay is undesirable in the vast majority of cases.

Improved PSPad's (editor) syntax and clip files in the Extras folder. [provided by GoGoGadgetScott & beardboy]

Improved SendMessage and PostMessage to support "ahk_id 0xFFFF", which is can be used to send HWND_BROADCAST messages to all windows in the system. [thanks Mythos]

Added a "fast" option to PixelSearch, which is up to 100 times faster for large search areas such as the entire screen. [code by Aurelian Maga]

Added command ImageSearch, which searches the screen for an image. [code by Aurelian Maga]

Added "WinGet Transparent" and "WinGet TransColor". Also added "WinSet Redraw", which refreshes a window by repainting it. [thanks Mythos]

Added the "WinSet Style|ExStyle" and "Control, Style|ExStyle" to change the styles of windows and controls not owned by the script.

CHANGES FOR GUI

Fixed "Gui +/-Disabled" and "Gui +/-AlwaysOnTop". In addition, it is no longer necessary to do "Gui Show" afterward. [thanks beardboy]

Fixed the inability to set a progress bar back to the default color. [thanks toralf]

Improved "Checked", "Disabled", and "Hidden" to recognize an optional 1 or 0 suffix. Example: Gui, Add, Checkbox, Disabled%Disabled?% Checked%Checked?%. [thanks MikeG]