Version 1.0.27
Changed A_OSVersion to recognize Windows Server 2003 (as WIN_2003).

Post/SendMessage can broadcast to ALL windows via ahk_id 0xFFFF, as intended. [thanks Mythos]

If you're using Post/SendMessage to trigger menu items created via the Menu command, each ID is now 10,000 higher than it was before.

Improved RunWait to support OutputVarPID (like the Run command).

Added built-in variables A_IsCompiled and A_IsSuspended, which contain 1 if the script is running as a compiled EXE or is suspended, respectively. [thanks Mythos]

Added commanded OutputDebug, which sends a string to the debugger (if any) for display. [thanks procyon]

CHANGES FOR GUI

Fixed windows containing tab controls: previously, controls outside the tab control were sometimes invisible when the window was first shown. [thanks ranomore]

Changed "GuiControl +/-Options" to set ErrorLevel to 1 (failure) if none of the specified styles are allowed to be applied.

Increased the maximum number of GUI windows from 10 to 99, and the maximum number of controls per window from 1000 to 11000.