Version 1.0.28
FileAppend's text parameter may be omitted to create an empty file or update the timestamp of an existing file. [thanks JSLover]

The following change from v1.0.26 was undone because it is not needed: "Changed hotkeys such as "a & b::" to be seen as "~a & b::" if "~a::" also exists as a hotkey." In such a case, the hotkey "~a::" causes an A keystroke to be sent whenever A wasn't held down to modify another key such as B.

Improved SoundPlay to play standard system sounds (hand, question, exclamation, and asterisk). [thanks Mythos]

Improved the tray icon to optionally be unaffected by Pause and Suspend. The icon can be "frozen" by specifying 1 as the last parameter of the Menu command. Example: Menu, Tray, Icon, C:\My Icon.ico, , 1

Added ahk_pid as a way to find window(s) matching a process identifier (PID). Example: WinClose ahk_pid %VarContainingPID% ; Closes only the first matching window.

Added command SoundBeep, which emits a tone from the PC speaker. [thanks Tekl]

Added support for hotkeys that fire upon release of a key. Examples:
*LWin::Send {LControl down}
*LWin up::Send {LControl up}
^!r up::MsgBox You pressed and released Ctrl+Alt+R.