Version 1.0.28.01
Improved FileAppend and file-reading loops to be able to write text to standard output (stdout) by using an asterisk as the filename. Although such text can be redirected to a file, piped to another EXE, or captured by fancy text editors, it will not appear at the command prompt it was launched from (there is a workaround described at FileAppend). For example, the following would be valid if typed at a command prompt:
"%ProgramFiles%\AutoHotkey\AutoHotkey.exe" "My Script.ahk" >"Error Log.txt"