Version 1.0.31
NOTE: Although this release has been extensively tested, quite a few low-level changes were made. It is recommended that mission-critical scripts be re-tested -- or that you wait two weeks for any bugs to get fixed -- prior to adopting this version.

Fixed the inability of a timer to reset itself. For example, if "SetTimer, MyTimer" is done within the MyTimer subroutine, the timer will begin waiting for an entirely new interval to expire before running again.

Fixed syntax checker reporting parentheses inside of quoted strings as unbalanced.

Fixed GuiControl to be able to change the ReadOnly attribute of an Edit control. [thanks Bahri]

Changed expressions so that references to invalid dynamic variables such as Array%VarContainingSpaces% produces a runtime warning.

Lowered the precedence of string concatenation in expressions. Although this is unlikely to affect any scripts, if you have any doubt please consult the operator precedence table.

Improved string concatenation in expressions to work on anything (e.g. arrays and sub-expressions). In addition, a period can be used as a concatenation operator to improve readability.

Dynamic variables such as Array%i% are up to 1000 times faster. A script may now have several million variables without a large slow-down. [thanks Laszlo]

Improved the performance of GuiControl and GuiControlGet when used on a window that has hundreds of controls.

Expressions that use a lot of ANDs and ORs are faster due to short-circuit boolean evaluation.

Improved memory conservation in scripts that use thousands of variables by adding two new variable capacities: 1 and 6. Formerly, the smallest capacity (other than zero) was 63.

Added built-in variables A_LineNumber and A_LineFile, which contain the currently executing line number and the file name to which it belongs. [thanks toralf]

Added directive #ClipboardTimeout, which changes how long the script persists in trying to open the clipboard (default 1000 ms). [thanks Tekl & Jason Payam Ahdoot]

Added the ability to pass parameters and receive return values from functions. Functions also support local variables. In addition, ListVars can show the contents of local variables to help debugging. See functions for details.